% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_pkg.R
\name{mk_package}
\alias{mk_package}
\title{Make Package}
\usage{
mk_package(
  path,
  author,
  email,
  git_username = NULL,
  git_email = NULL,
  readme_md = TRUE,
  git = TRUE,
  check_pkg_name = TRUE,
  ci = "travis",
  license = "MIT",
  pkgdown = TRUE
)
}
\arguments{
\item{path}{DESCRIPTION.}

\item{author}{DESCRIPTION.}

\item{email}{DESCRIPTION.}

\item{git_username}{DESCRIPTION.}

\item{readme_md}{DESCRIPTION.}

\item{git}{DESCRIPTION.}

\item{check_pkg_name}{DESCRIPTION.}

\item{ci}{DESCRIPTION.}

\item{license}{DESCRIPTION.}

\item{pkgdown}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
