% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_pkg.R
\name{mk_pkg}
\alias{mk_pkg}
\title{Make Package}
\usage{
mk_pkg(
  path,
  author,
  email,
  git = TRUE,
  git_username = NULL,
  git_email = NULL,
  readme_md = TRUE,
  check_pkg_name = TRUE,
  license = "MIT",
  pkgdown = TRUE
)
}
\arguments{
\item{path}{character. The path to create the package at.}

\item{author}{character. The name of the package author.}

\item{email}{character. The email to use for Git and GitHub.}

\item{git}{logical. Whether to initialize a Git repository for the package.}

\item{git_username}{character. The username to use for Git and GitHub.}

\item{git_email}{character. The email to use for Git and GitHub.}

\item{readme_md}{logical. Whether to create a README file in markdown format.}

\item{check_pkg_name}{logical. Whether to check that the package name is available on CRAN.}

\item{license}{character. The license to use for the package (either "MIT" or "GPL-3").}

\item{pkgdown}{logical. Whether to create a pkgdown site for the package.}
}
\description{
Create a new R package with the specified parameters
}
