% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_config.R
\name{import_config}
\alias{import_config}
\title{Package Config}
\usage{
import_config(file = NULL, file_type = "yaml")
}
\arguments{
\item{file_type}{DESCRIPTION.}

\item{path}{DESCRIPTION.}
}
\description{
FUNCTION_DESCRIPTION
}
