% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_config.R
\name{import_config}
\alias{import_config}
\title{Package Config}
\usage{
import_config(path, file_type = "yaml")
}
\arguments{
\item{path}{character. The path to the configuration file.}

\item{file_type}{character. The file type of the configuration file (either "yaml" or "json").}
}
\value{
A list containing the configuration data.
}
\description{
Import configuration data from a file.
}
