% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_pkg.R
\name{mk_pkg_from_config}
\alias{mk_pkg_from_config}
\title{Make Package from Config}
\usage{
mk_pkg_from_config(config_path, file_type = "yaml")
}
\arguments{
\item{config_path}{character. The path to the configuration file.}

\item{file_type}{character. The file type of the configuration file (either "yaml" or "json").}
}
\description{
Create a package based on a configuration file.
}
