% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_config.R
\name{write_config}
\alias{write_config}
\title{Write Config}
\usage{
write_config(path, config_data)
}
\arguments{
\item{path}{character. The path to write the configuration file to.}

\item{config_data}{list. The configuration data to write to the file.}
}
\description{
Write configuration data to a file.
}
